% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinRuler.R
\name{match_multi}
\alias{match_multi}
\title{Extension of the base 'match' function}
\usage{
match_multi(x, y, sep = ";")
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}

\item{sep}{character separating different items to be matched in a given y element}
}
\value{
a numeric vector with matching indices
}
\description{
Extension of the base 'match' function to the case where 
elements of y can contain multiple items to be matched
}
\examples{
x <- c("a", "b")
y <- c( "bb", "c;b", "aa", "c;a;b")
match_multi(x=x, y=y, sep = ";")
}
