% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinRuler.R
\name{proteinRuler}
\alias{proteinRuler}
\title{Compute protein abundance using the protein ruler methodology}
\usage{
proteinRuler(df, col_names = "Gene names",
  col_protein_id = "Protein IDs", col_score = "Score",
  Score_threshold = 0, pattern_intensity = "^Intensity.",
  col_intensity = NULL, organism = "mouse",
  mass_per_cell_in_pg = NULL, DNA_mass_per_cell = 5.5209e-12, ...)
}
\arguments{
\item{df}{A data.frame containing protein intensities. By default, protein intensity column names start by "Intensity." 
(use parameter \code{pattern_intensity} to change)}

\item{col_names}{Column with gene names.}

\item{col_protein_id}{Column with protein IDs}

\item{col_score}{Column with protein identification score}

\item{Score_threshold}{Threshold on protein identification score}

\item{pattern_intensity}{Pattern (regular exrpression) used to identfy df's columns containing protein intensity values}

\item{col_intensity}{Names of intensity columns. Overrides \code{pattern_intensity}.}

\item{organism}{Either "mouse" or "human"}

\item{mass_per_cell_in_pg}{Compute protein abundance using a constant mass per cell.}

\item{DNA_mass_per_cell}{Mass of DNA per cell (in g)}
}
\value{
a data.frame with protein abundances
}
\description{
Compute protein abundance using the protein ruler methodology
}
